i tried everything
The broken buildstate only sets the visualizer renderer to enabled and sets the linked gameobjects renderers to false.
i even made a broken buildstate for each child and assigned the visualizer hoping it would be a temporary solution but even that didn't work.

i've been writing scripts that attempt to enable the meshrenderers & colliders on the children when CurrentBuildIndex = -1

or if isBroken = true..

i subscribed to the OnBuildState event and sent a check everytime the event triggered to check if BuildStateIndex = -1. if BuildStateIndex < 0 then foreach meshrenderer and collider in BrokenState.GetComponenetsInChildren<Meshrenderer/Collider>(true){
meshrend.enabled = true..
}
you get the point..

short summary of what happened, i've been at this since yesterday morning hopefully you can figure it out because i am on the verge of smashing my head into a wall xD

as a last option custom buildstate manager lol