Basic tower setup | Changed faction name | Faction Lore is complete | (Read commit desc)
Tower:
Added on off switch functionality
Added Setting functionality
RequiredPower will scale up based on Setting. (Setting * 15).
Added Mode, Lock & Error states, currently only lock is the only thing that does something.
Fixed cable connection visualizer pointing down.
Fixed cable connection only connecting if cable was a corner pointing up.
Towers are added to a list (AllTowers)
DefaultSignal is set as Setting on Awake, (Default tower signal). Removed on destroyed.
Moved ToolAssigners to their own class.

Assigned can logic read & write.

Tower Setting is now Serialized and saved, which ensures setting is not lost on reloads.
We also handle setting change on dedicated (I think).

Setting is clamped to 0 and 1000

Tower build states work normally with the only issue being that the interactable visualizer when you hold a tool to construct or deconstruct the tower seems to be adding the broken state to that visualizer making it bigger than the base tower. hopefully that made sense.

I mentioned on discord the big yellow box, if the mesh renderer of the broken state is removed from the broken state the box is normal and is the correct size.

tried to fix it with a few options but failed. the only thing that slightly worked was when the broken state is disabled completely in unity. this makes the yellow box the correct size but the only issue is that when the tower is broken the color of the broken state doesn't care about the color of its parent. (Spawns white), could maybe set color of broken state in code probably, not sure.

Please don't hesitate from removing code we don't need 😘